Add resizing filters for CONFIG_DSPL_RESIDUAL
When downsampling, filters are necessary to remove high frequency
components to prevent aliasing.

For this experiment, the Lanczos3 filters will be used for both down and
up sampling. These filters have already been implemented in the
experimental branch but are not available in the master branch.

This commit adds these filters.

Relevant CL from the experimental branch:
https://aomedia-review.googlesource.com/c/aom/+/115027

Change-Id: Ief554f9a45d51cd2beb2336f6dcb4238036430d0