Modify bitstream for CONFIG_DSPL_RESIDUAL
This change writes the encoder decision w.r.t. downsampling to the
bitstream. Specifically, for every partition block for which:
    i) does not skip transform (mbmi->skip_txfm != 0)
   ii) has sides bigger than DSPL_MIN_PARTITION_SIDE
  iii) is an inter block

the encoder will write the downsampling decision (mbmi->dspl_type) to the
bitstream. This decision is written (if applicable) right after
mbmi->skip_txfm is written.

A new CDF (ec_ctx->dspl_type_cdf) is introduced to write this decision with
entropy coding. Additionally, the costs for writing this decision is
stored in x->mode_costs->dspl_type_cost.

The probability optimizer tool in tools/aom_entropy_optimizer.c is also
updated to support this new CDF.

Change-Id: I47eb1d9f219c726346cd6067940c3fd5c533e6d9