Initialize dspl_type for CONFIG_DSPL_RESIDUAL
The coded will store the downsampling decision in mbmi->dspl_type (which
is read by the forward and inverse transform functions).

However, these functions are part of many codepaths and thus it is
critial to initialize mbmi->dspl_type correctly.

This change does this initialization, setting mbmi->dspl_type to
DSPL_NONE.

Change-Id: Id897275cfc845a08d889bdfed61ff98137ad0977