Adds an optional refresh parameter to subgop cfg
A new refresh parameter is added to subgop cfg.
The refresh parameter provides a hint on the level
of the reference frame to refresh for each subgop
step. If the specified level is not provided, the
value in the refresh parameter will be 0. An encoder
is expected to honor the hint if a reference of the
provided level exists in the reference buffer list.
If a reference of the specified level does not exist,
the encoder can decide the buffer to refresh on its
own.

Change-Id: I9db3d77aa529a63a43e48fa4b865ac332e91a6d0