Extend instead of copy, for blocks on frame edge
When building the intra-predictor border, it is possible that a block
extends beyond the frame edge, e.g.:

--------------------
IMAGE              |
                   |
                 ++++++++
                 +Block +
---------------- +      +
                 ++++++++

In these cases, some of the intra-predictor pixels to the left
of the block are valid, whiles others are not. Same for top.
Detect this case and if invalid pixels are detected, extend
existing values instead of copying over the invalid ones.

Change-Id: I48cbf0f96230f1c362c7f9f8cc06652dc9c61708