Introduce experiment CONFIG_DSPL_RESIDUAL
In this experiment, a new partition level downsampling based
coding tool is explored.

For each partition block, the encoder can either code the residuals as
is, or they can be coded after downsampling them by a factor of 2 in
both directions.

This decision by the encoder will be signalled in the bitstream.

The code for this experiment will be wrapped in compiler flag
CONFIG_DSPL_RESIDUAL.

Change-Id: Ic9b399490eaf5fd325fd6850d599cd2ef72b312b