Modify bitstream for CONFIG_DSPL_RESIDUAL
This change writes the encoder decision w.r.t. downsampling residual to
the bitstream. Specifically, for every partition block which:
    i) does not skip transform (mbmi->skip_txfm != 1)
   ii) has sides bigger than DSPL_MIN_PARTITION_SIDE
  iii) is an inter block

the encoder will write the downsampling decision (mbmi->dspl_type) to
the bitstream right after is_inter is written.

A new CDF is introduced to write this decision with entropy coding.
Additionally, the costs for writing this decision are stored in
x->dspl_type_cost.

The probability optimizer tool in tools/aom_entropy_optimizer.c is also
updated to support this new CDF.

Change-Id: I6437499a6769f2ddcd23f66388763b8378d5a415