Separate txsearch hashing for CONFIG_DSPL_RESIDUAL
Due to overlap of the partition and transform searches, the encoder
hashes transform blocks and does the transform search iff the selected
transform block residual is unseen.

With the introduction of a new downsampling option, these results need to be
separately maintained for either downsampling option.

This change makes this modification by add a new dimension to the hash
tables to store the results for either downsampling option.

Change-Id: Ief1fe17a7d052375ef46ccfee76c8be42f694942