Modify quantizer for CONFIG_DSPL_RESIDUAL
Downsampling is a lossy operation which significantly reduces the rate
cost for coding the transform coefficients. To offset this, it is
necessary to modify the quantizer to be less aggressive when
downsampling.

This change makes this modification. A new method setup_dspl_quantizer
is introduced to modify the Y-plane's quantizer by a given qindex
offset. Note that this method merely swaps pointers to already built
quantizers, the quantizers are not re-built.

On the decoder side, the quantizers are built only for the required
qindex(s). Thus new quantizers are built for the downsampling option.

The qindex offset is currently static and is set by the calc_dspl_qindex
method.

Change-Id: I02dce5d69391aae0a5b16071e8074931393569fd