Modify transforms for CONFIG_DSPL_DEBUG
Depending on encoder decision, some partition blocks need to have their
residuals downsampled before coding (and upsampled before being added to
prediction).

This change modifies the forward and inverse transform functions to do
this operation. Note that the transforms are still signalled at their
original size.

Specifically, mbmi->dspl_type controls the transform behaviour. If
mbmi->dspl_type is set to DSPL_XY, the forward transform
av1_xform_quant() will downsample the residual, perform a smaller size
transform operation and pack the coefficients in the coeff buffer.

Since the downsmapled coefficients only amount to 1/4th of original
coefficients but are packed in the full size coeff buffer of the
original transform size, the coefficients are "packed" in such a manner
so as to ensure that the quantizer sees consecutive zeros for the
remaining 3/4th of the coefficients.

Similarly, if mbmi->dspl_type is set to DSPL_XY, the inverse transform
functions unpacks, performs the inverse transform, upscales and adds the
result to the prediction.

If mbmi->dspl_type is set to DSPL_NONE, the transform behaviour is
unchanged.

A new method av1_inverse_transform_diff is introduced to extract the
residual in a 16-bit signed integer buffer. This is necessary because
the inverse transform method itself has a limited range (depending on
bit depth) and always outputs to an unsigned buffer.

Change-Id: I54462a475e74156b3414b34ae830d0fbf072e608