Update superblock level RDO for CONFIG_DSPL_RESIDUAL
In AV1, the coding context of nearby (primarily the above and left)
blocks is used to better estimate the probability distrubution for
encoding the current block.

In experiments, it was noticed that while partition level RDO was
locally optimal, on some videos like ice_cif.y4m it leads to regressions
beyond what can be explained using the cost of addition signalling of
the downsampling option.

To address this, this change introduces RDO at the superblock level.
Specifically, if the RD cost with the downsampling enabled is
higher than that with downsampling disabled, all partition in that
superblock are coded without downsampling.

This yields marginal improvements.

Change-Id: Ibe1b2fac4507463daec5cc3b850fd7ae7fa2ddc4