Type-fix for inter_tx_size
inter_tx_size should be of type TX_SIZE, but it is listed as uint8_t.
Although this is currently fine (the UENUM1BYTE macro maps the enum
to uint8_t), it should really be TX_SIZE, so if the enum definition
ever changes, this data structure will be automatically updated.

Change-Id: I1426071224ae5cf7c8328fe172905c0179cc73de