Fix crash when all past frames are level 1
The refresh rule was not set up to be able to refresh level 1
frames unless we are also coding a level 1 frame. In some
subgop configurations, it may be possible to have many level 1
frames and be forced to refresh one of them. In that case, we refresh
the oldest one.

Change-Id: Id25744766fd8242afd524259f55932c4dfd57e15