Fix high energy detection threshold in HBD encode
The threshold for detecting high residual energy blocks is corrected for
high bit-depth encoding. The bit-depth based scaling of the threshold is
not needed since the sse is down-converted to 8-bit domain.

Results on 10-bit encode show minor BD rate loss (0.03%) for speed 0,
and more or less neutral BD rate impact for other presets.

STATS_CHANGED for hbd encoding.

Change-Id: Ic8449ef8d9230fa8982b2b01eae999db8cb96d60