LOOP_RESTORE_CNN: Disable CDEF for RESTORE_CNN
CNN models are trained to work on pre-CDEF input, so we disallow this
combination.

For now, this includes a 2nd pass of CDEF + LR filters if this combo was
picked in the first pass, which could be improved later.

Brings LOOP_RESTORE_CNN even closer to CNN_RESTORATION.

All Intra
---------
                  CNN_RESTORATION   LOOP_RESTORE_CNN
lowres2           -4.701            -4.485
midres2           -5.083            -4.797
objective-1-fast  -4.865            -4.500

Random Access
-------------
Pending

Change-Id: Ic8c8ea414aa89ddc4f92ba9a62f2d10f0856f5ea