Further Lambda scaling tweaks.
Change I5592b2aa adjusted the scaling of the rd lambda to be based
on layer depth not frame type.

Some gains were reported for low res and midres sets in ssim though
the results were mixed in opsnr.

However there was a significant regression in rate accuracy.

This change tweaks the layer numbers for ARFs a little further and gets
some additional metrics gains of ~0.2 to 0.25% (lowres2 and ugc360) up
to -0.4 to -0.55% for midres2 over YT rate ranges. To counter balance
the impact of these changes on rate there is also a small change to the
calculation used to define n allowed Q range.  In effect the RD lambda
will tend to reduce the number of bits spent for a given Q by reducing
bits spent in the prediction signal. This is countered by lowering Q
slightly which will shift the bit allocation balance slightly in favor
of the residual signal.

Some refactoring and elimination of duplication.

STATS_CHANGED

Change-Id: Ie1cc0f911cb9f413f7ac7e41d829e8e3cfef61f9