Low-delay config: No level 1 frames in last subgop
Specifically, use level 4 instead of 1 for frame 14 in last subgop, as
this frame will be used as reference for very few frames after it.

Also, change the ratecontrol code to support this case, where no frames
in a subgop are at level 1.

Note: This also fixes the qp allocation of the last frame, if it's the
only frame left in the last subgop. Earlier, this got same qp as level
1; now we assign qp based on the actual layer depth of this frame in the
subgop config.

BDRate (Ovr PSNR) Before vs After:

- CTC with --use-fixed-qp-offsets=1:

lowres2:          -2.733
midres2:          -2.740
objective-1-fast: -2.798

- CTC with --use-fixed-qp-offsets=0:

lowres2:          -2.367
midres2:          -2.523
objective-1-fast: -2.578

BUG=aomedia:2832

Change-Id: I42640af9a2ee7245798b9c84f70aff04e3f81bb4