Reduce memory used for MACROBLOCK::mbmi_ext
The mbmi_ext struct in MACROBLOCK was previously an dynamically
allocated array whose size is the number of mi's in a superblock.
However, only the first element of the array was ever used.

This commit change the member struct from a dynamically allocated array
to a simple struct allocated on the stack.

This change reduces the memory consumption by 2.8% on a lowres clip.

Change-Id: I61a33577f0ff6ef63ce883bb9c46fb3efb48a8e8