Reuse prediction mode in EXT_RECUR_PARTITIONS
This commit adds the option to reuse the prediction mode from earlier
partition search. The option can be turn on by setting
USE_OLD_PREDICTION_MODE macro in av1/common/enums.h to 1. Due to an
issue to spatial context, currently this only enables mode reuse if the
cached mode is an intermode.

Performance on Speed 2 in q mode:
 TEST_SET | BDRate (Exp) | Speed (Exp) | BDRate (Base) | Speed (Base)
  MIDRES  |    +0.3%     |    +39%     |     -3.3%     |    -518%
  LOWRES  |    +0.4%     |    +41%     |     -2.8%     |    -490%

Change-Id: I956d89f56af993ceb67ddc184afec38c0df775ed