Fix colors count and bias threshold in hbd path for palette.
This patch fixes the colors count which is used internally to populate
the dominant colors array in av1_rd_pick_palette_intra_sby(). The count
is calculated based on the original hbd colors in the block. Also the
threshold used for bias in optimize_palette_colors() is modified based
on the bit_depth.

STATS_CHANGED for hbd encoding

Change-Id: I18ff6c1c67dc4b9e15346f6a10983140c247865c