Correct av1_encoder_get_relative_dist()
av1_encoder_get_relative_dist() is used only on the encoder side to make
encoding decisions such as prediction mode selection. It does not need
to check if order hint is enabled in the bitstream.

This fixes some incorrect encoder behaviors in some corner cases. For
example, before this fix, with --enable-order-hint=0
--enable-onesided-comp=0, the encoder would skip all compound modes.

Change-Id: Ifacfa67accff55eb4bb161025c3c5a05f9dde475