Decouple luma from chroma during intra mode search
In libaom, the intra mode search in interframe used to work as the
following:

>> for each luma intra mode:
>>   search luma intra mode
>>   if chroma intra search is done:
>>     retrieve chroma mode from cache
>>   else:
>>     search all chroma intra modes
>>     store chroma intra mode in cache
>>
>>   merge luma and chroma stats

A perk of this workflow is that the rdcost of CFL is always calculated
using the first luma mode searched. In this CL, we instead rework the
search to the following:

>> for each luma intra mode:
>>   search luma intra mode
>>
>> if best luma intra mode is better than best inter luma stats:
>>   search chroma intra modes
>>   merge stats

The new search process is easier to understand and mirrors the process
in intraframe coding. Furthermore, this gives decent performance gain on
hdres testset while maintaining the current speed.

Performance:
 SPD_SET | TESTSET | AVG_PSNR | OVR_PSNR | SSIM | SPD
 TBA

STATS_CHANGED

BUG=aomedia:2863

Change-Id: I89e2c6425cf1b56f8a946a954c79ab413167323f