Decouple luma from chroma during intra mode search
In libaom, the intra mode search in interframe used to work as the
following:

>> for each luma intra mode:
>>   search luma intra mode
>>   if chroma intra search is done:
>>     retrieve chroma mode from cache
>>   else:
>>     search all chroma intra modes
>>     store chroma intra mode in cache
>>
>>   merge luma and chroma stats

A perk of this workflow is that the rdcost of CFL is always calculated
using the first luma mode searched. In this CL, we instead rework the
search to the following:

>> for each luma intra mode:
>>   search luma intra mode
>>
>> if best luma intra mode is better than best inter luma stats:
>>   search chroma intra modes
>>   merge stats

The new search process is easier to understand and mirrors the process
in intraframe coding. Furthermore, this gives decent coding gain
with fairly small speed loss on lower speed setting. On high speed
setting, there is a slight speed up with some quality loss due to
interaction with winner mode.

Performance in vbr mode over 70 frames:
 SPD_SET | TESTSET | AVG_PSNR | OVR_PSNR |   SSIM  |  SPD
---------+---------+----------+----------+---------+-------
         |   HDRES |  -0.095% |  -0.086% | -0.088% | -0.4%
    1    |  MIDRES |  -0.100% |  -0.131% | -0.164% | -0.8%
         |  LOWRES |  -0.044% |  -0.036% | -0.101% | -0.2%
---------+---------+----------+----------+---------+-------
         |   HDRES |  -0.071% |  -0.070% | -0.087% | -0.1%
    3    |  MIDRES |  -0.090% |  -0.084% | -0.053% | -0.6%
         |  LOWRES |  -0.049% |  -0.042% | -0.073% | -0.1%
---------+---------+----------+----------+---------+-------
         |   HDRES |  +0.003% |  -0.006% | +0.007% | +0.5%
    4    |  MIDRES |  +0.067% |  +0.067% | +0.097% | +0.1%
         |  LOWRES |  +0.016% |  +0.022% | +0.003% | +0.4%
---------+---------+----------+----------+---------+-------

STATS_CHANGED

BUG=aomedia:2863

Change-Id: I89e2c6425cf1b56f8a946a954c79ab413167323f