Don't enable CONFIG_DEBUG for RelWithDebInfo
Don't enable CONFIG_DEBUG if CMAKE_BUILD_TYPE is RelWithDebInfo. Only
enable CONFIG_DEBUG if CMAKE_BUILD_TYPE is Debug.

The code enabled by CONFIG_DEBUG is debugging code and therefore it may
change the behavior of the binary. The goal of the RelWithDebInfo build
type is to add debugger symbol information to a release build, so it
should not enable debugging code. (For example, RelWithDebInfo disables
assertions by -DNDEBUG.)

BUG=aomedia:2866

Change-Id: Ib867ec689431934ddfddade526454fd761921e3c