Allow insymmetric frame buffers for temporal filtering.
This patch changes how frame buffers are chosen for temporal filtering.

First, it allows temporal filtering to use different number of preceding
and succeeding frames. This allows for the possibility of using more
frames when the sequence is stable to better remove noise.

Second, before this patch, noise level is used to determine additional
frames to use for temporal filtering. The argument was that when noise
is high, MV tends to get unreliable, and hence we should not use as many
frames. However, this also is in contradiction with the intuition that
when noise is high, we also need more frames to de-noise.

This patch uses the frame-level correlation derived from first-pass
stats to determine the number of frames used (how insymmetric can the
frames be, relative to the to-be-filtered frame). The derived
correlation is the correlation of the underlying random process without
addtional noise, and therefore should measure the stablity of MVs
better.

Coding gains was observed:

Speed1:
            avgPSNR    ovrPSNR    SSIM      vmaf
lowres      -0.091     -0.198     -0.118    -0.686
midres      -0.167     -0.251     -0.269    -0.361
ugc360p     -0.169     -0.260     -0.358    -0.108

Speed4:
            avgPSNR    ovrPSNR    SSIM      vmaf
lowres      -0.198     -0.323     -0.184    -0.655
midres      -0.095     -0.178     -0.101    -0.143
ugc360p     -0.223     -0.293     -0.440    -0.122

STATS_CHANGED

Change-Id: I53edd0efacba949563dbce17c732c9d99e633cfe