Modify AVxEncoderThreadTest to test temporal filter
This CL enables temporal filtering in AVxEncoderThreadTest and
AVxEncoderThreadTestLarge. The number of frames being tested is 6.
The value of arnr_maxframes is set to 5 to enable KF filtering.
Also, g_lag_in_frames is set to 7 with an additional parameter
max_gf_interval being set to 4 in order to enable ARF filtering.

Change-Id: Ia21041ffc76b54260442cb064c21b7f31f19a728