Relax range constraints for MV adjustment
Remove range constraints on MV adjustment because they not necessary
once MVs are refined for subblocks. MV refinement quantities (vx, vy)
are also scaled by a factor of 2, which provides the best quality
improvement.

This change brings about 0.05% gain in CONFIG_OPTFLOW_REFINEMENT,
and additional 0.6-0.7% gain when USE_OF_NXN is on (i.e., MV
refinement is applied to subblocks).

Change-Id: I827f12c3ffb21edd8f249e996c2be514d8eda462