Further tweaks to rate adaption.
This is a further experimental change to the rate adaptation.
The rate error factor is based on the bits off target error
measured against the larger of the actual bits used so far and
the projected number of bits left. This tends to naturally damp
the factor so that it is likely to be less extreme at the two ends
of the clip.

A larger adjustment is allowed at the end of the first GOP.
The specified undershoot and overshoot limits are used to
define the maximum amount of adjustment allowed.

This gives some further improvements in metrics along with
a small improvement in average rate accuracy over the previous
change. (Overall psnr, ssim, abs_rcerror (-ver better)

low res:  -0.076  -0.009  -0.462
ugc360: -0.071  -0.043  -0.643
midres2 -0.029  0.064  -0.783

STATS_CHANGED

Change-Id: Ia7d2bc833a4266ec92e9c4e78caa4d80ed72f3af