Further tweaks to rate adaption.
This is a further experimental change to the rate adaptation.
The rate error factor is based on the bits off target error
measured against the larger of the actual bits used so far and
the projected number of bits left. This tends to naturally damp
the factor so that it is likely to be less extreme at the two ends
of the clip.

A larger adjustment is allowed at the end of the first GOP.
The specified undershoot and overshoot limits are used to
define the maximum amount of adjustment allowed.

This gives some further improvements in metrics when
overshoot and undershoot are set to 100 along with a small
improvement in average rate accuracy.

WIP : Figures pending and also for OSUS 25

STATS_CHANGED

Change-Id: Ia7d2bc833a4266ec92e9c4e78caa4d80ed72f3af