Improve border ext support for lanczos resampling
Changes symmetric to mean mirror half where the end pixels are
repeated. Additionally, reflect padding and gradient preserving
padding are also added.
The gradient padding implemented is not exactly the text book
method, but can be improved.

The order of efficiency of down-up operations in best to worst
order is:

Gradient
Symmetric
Repeat
Reflect

Change-Id: I40d14c541aef00e8d6be1768c578e0e56fb8fbd1