Adjust keyframe boost for screen content
Do not take accumulated second ref error into kf boost calculation.
This variable is suppose to measure the gradual change and error
accumulation. This is often true for natural videos where scenes
are changing gradually.

For screen content, the change is often abrupt, and the smooth
transition effect is not usual.

Gains on speed 1, 150 frames:
-0.20% ovr_psnr, -0.46% ssim, -0.20% vmaf_neg.

BUG=aomedia:2650

STATS_CHANGED

Change-Id: I0d5f214cd7a2b7b379e235f0c4ffa51a3f201d46