Add bit per frame to indicate merge.
Currently in CONFIG_RST_MERGECOEFFS, an additional bit is written for
each unit indicating whether that unit is merged with the previous. Now
we write an additional bit for each frame to indicate if there are any
merges - if not, the merge bit for each unit is not written/read.

Change-Id: I347d050c753cb252936170375b4d7c8931620653