Create new reference mapping for syntax changes
This adds a structure to contain a new mapping based on a
ranking. The ranking is currently based only on display order
distance from the current frame, but will be updated in the future
to also use frame quality. This new mapping will help to removed
the named reference scheme. No change in performance.

Change-Id: I0286129538da504dab3dea5554c56270a41741c1