ERP: Add is_coded_map to record blocks' status
Since it will be hard to figure out an explicit algorithm for
checking whether the top right block is available once flexible
partitions are enabled, we maintain an array is_mi_coded to record
at the superblock level whether an mi(in 4x4 granularity) has been
coded. The map is maintained separately at decoder, encoder, and in
RDO.

Note: There is a STATS_CHANGED when this is run with TPL model because
the coded map is not set for TPL model. This however does not affect the
results under CTC.

Change-Id: Ic4c878bd59dd05c6c68762bf66213a142f19448a
(cherry picked from commit 0a044385e7bd84935160357d6284c14ff9b14292)