Opfl refinement without second motion compensation
Introduce a framework without second pass of motion compensation
in optical flow MV refinement. It is enabled when the macro
OPFL_SECOND_PASS_MC is set to 0.

In this framework, the compound predicted block is refined
using the MV offsets and the gradients. It does not require a
second motion compensation and hence reduce the decoder complexity.
The original 2 MC framework has been refactored for this change.

On lowres2
                    BD-Rate   Enc time   Dec time
2 MC (original)     -3.180    +15.409%   +45.973%
1 MC (this change)  -2.371    +19.377%   +38.674%

Change-Id: Ieec9689174e38affa8b62dd43b34f3ccfd63dae6