Consider pcnt_neutral in key frame detection
Add a new contraint such that a frame is considered as a key frame
when the frame is not very dark (weight measures the brightness of
a frame, the darker a frame is, the higher weight we get), and
its pcnt_neutral is below a certain threshold.

This restores the quality regression on a gaming clip where the
frame is in transition and an unnecessary key frame is inserted.

STATS_CHANGED

BUG=aomedia:2926

Change-Id: I847f4323a2fb665de9283a85d72f2b677da47612