Refactor rd loop for NEW_REF_SIGNALING
Iterate over all modes and reference combinations rather
than using the lookup table of pre-determined combinations.
It includes a function used to skip compound modes that
are currently prohibited by the bitstream syntax. This
is temporary and will be removed once the syntax is updated
to allow the more flexible reference scheme.

Change-Id: I46702bca04668568f3ba4541c19d9839f75d8402