Add utility/example to do CNN restoration on y4m
This utility is meant to be applied to a y4m input to
apply restoration on each frame using one of 4 supported models.
The four models supported are ones that are meant to
be used after down-up resampling by factors 2/1, 3/2, 4/3 and 5/4.

As an example, after 3/2 down-up resampling without compression,
the output PSNR on the CrowdRun video improves by about 1.4dB.

Use -DCONFIG_TENSORFLOW_LITE=1 with cmake to build this utility
in libaom.

Change-Id: I53446c269161d223787b83995fcc52c4e430c38e