Move an mv_cost table init from temporal filter to tpl_model
Since temporal filter uses sad instead of entropy for mv_cost, the
initialization should not be done in temporal filtering. This commit
moves it to the correct location.

Furthermore, this fixes an incorrect behavior where temporal dependency
model would use 0 as the mvcost in motion estimation. This happens in
the rare scenario where there is not enough frames in the kf group for
temporal filtering, so the mv_cost is not initialized.

There are some small changes on a couple clips, but the overall
performance is neutral.

CHANGES_OUTSIDE_COMPILER_FLAG
STATS_CHANGED

Change-Id: I90450f1095348bc9033c57144c3dd74783ae197e