FLEX_MVRES: Add syntax for sb-level precision
This commit introduces the compiler flag CONFIG_FLEX_MVRES for the new
experiment.

As a first step, this adds use_sb_mv_precision under FeatureFlags
struct.

The variable use_sb_mv_precision is sent through the bitstream at the
beginning of each from provided that the subpel precision is on.
Then for each superblock of the frame, if use_sb_mv_precision is on, the
encoder additionally sends the entropy-coded sb_mv_precision.

This commit only adds the plumbing needed to send the flags, the actual
mv precisions are not modified by this change.

BUG=aomedia:2958

Change-Id: I2fbcc53964f77865febc2f21a76697982a898189