Remove obsolete code related to superframe index
The MAG_SIZE macro and index_size local variable were added in the
following CL related the writing of the superframe index:
https://aomedia-review.googlesource.com/c/aom/+/17140

Remove the ctx->pending_frame_sizes[] array, which was only used by the
removed write_superframe_index() function. The related
ctx->pending_frame_count field was also only used by
write_superframe_index(), but it has since found a new use in the
current code, so it is kept.

The writing of superframe index was first removed (conditionally if
CONFIG_OBU was equal to 1) in the following CL:
https://aomedia-review.googlesource.com/c/aom/+/24460

I did not track down which CL removed the write_superframe_index()
function completely.

Change-Id: I937043f438c46f93459ca9c310adf2bcb40d9531