rtc: Delay top-right wait in row-mt for rtc mode
In row-mt multi-threading, a thread waits for the
encoding of the top-right superblock to be complete
before encoding the current superblock. This wait
time is reduced in rtc mode, by delaying the waiting
of threads until the need for the data from the
top-right superblock region, i.e., when the right
edge of the block is also the superblock egde.

Change-Id: Ibf9d50bd1b4032da39c2ca6e2bab287db7166fad