Do not skip motion search even if zero mv error is small
Remove the condition that depends on a hard-coded threshold
to skip motion search in the first pass.

The removal should benefit videos with low texture and
those of high temporal correlations. The gain
on ugc480p and screen content set supports the hypothesis.

Results: speed 2, 150 frames
neutral on lowres and ugc360p.
-0.2% gain on ugc480p
-0.5% gain on screen content set

STATS_CHANGED

Change-Id: I664ce6ab0f8cb22943b336686d1b706f3ce93cf7