Allocate buffers related to last frame conditionally
Previous frame buffer in lookahead queue and cpi->scaled_last_source
are not required for all-intra encode. Hence the same are not
allocated for all-intra encode. Added a check key_freq_max==0,
to code all the video frames as intra frames in first-pass or
lookahead stage.

BUG=aomedia:2967

STATS_CHANGED for good preset with kf-max-dist=0

Change-Id: I2e30e84d422b7f3793965107c6e090d4fb6adf54