Partition search with given order
We add a new functionality which allows the encoder to do
mode search for a given partition tree for each superblock.

We also adds a function to  write the partition tree of
each superblock into a text file, serving as the baseline.

We allow the encoder to read multiple text files that contain
partition configurations.

Encode each superblock with these partition configurations and
pick the one with the smallest rdcost.

Now this functionality is under "CONFIG_PARTITION_SEARCH_ORDER",
which is turned off by default.

Change-Id: I43dcb70ffea51f72c749ea33f36d475e3a85d1f2