Partition search with given order
We add a new functionality which allows the encoder to do
mode search for a given partition tree for each superblock.

We also adds a function to  write the partition tree of
each superblock into a text file, serving as the baseline.

Then we read the parition tree from text files to let the
encoder do mode search for this given partition tree.

Now this functionality is guarded by a macro "PARTITION_SEARCH_ORDER"
which is turned off by default.

Change-Id: I43dcb70ffea51f72c749ea33f36d475e3a85d1f2