Define ColorCost in the same way as MapCdf
The MapCdf and ColorCost typedefs are both pointers to three-dimensional
arrays. The reason ColorCost is defined as a pointer to a
four-dimensinal array is apparently to allow it to point to a const
array. We can solve that problem by defining ColorCost to be a pointer
to a const array.

A downside is that the new definition of ColorCost does not show the first
dimension, so I added comments to note the first dimensions of the
three-dimensional arrays that MapCdf and ColorCost point to.

Change-Id: I115d892f05b4af2208ab99ca743562e2100ab836