Change the first parameter of av1_fill_mv_costs()
The first parameter of av1_fill_mv_costs() is a pointer to
FRAME_CONTEXT, which is a large struct. Since av1_fill_mv_costs() only
uses the nmvc member of the FRAME_CONTEXT struct, change the first
parameter of av1_fill_mv_costs() to a pointer to nmv_context. This helps
code analysis.

Change-Id: Ife5884259eb47c7a8da66c300ac68c9ad3fc1dc3