Remove a duplicate computation of haar wavelet in firstpass
When the function av1_haar_ac_sad_8x8_uint8_input is called by
firstpass_intra_prediction, it is assumed that the block size is 16X16.
So harr is computed on the 16X16 block by calling the function 4 times.

When the firstpass block size is set to BLOCK_8X8 for screencontents,
this leads to 3 redundant calls.

This commit takes the actual firstpass block size into account to reduce
the complexity.

On speed 6 screen_contents set, we observe an overall speed up of 3.3%.

Change-Id: I8070018caf48b310291770e82aa09af3c74892e1