Allow updates of motion costs for intrabc
This commits enables the encoder to update the motion vector table for
intrabc at various resolutions. By default, this is turned on for good
mode, but is disabled for realtime mode and allintra mode.

The IntraBCMVCosts struct is moved from AV1_COMP to MACROBLOCK to
support multithreading.

Performance on screen_content set:

 SPD_SET | AVG_PSNR | OVR_PSNR |   SSIM  |  SPD
    1    |  -0.166% |  -0.209% | -0.107% | +0.1%
    2    |  -0.284% |  -0.298% | -0.415% | +0.0%
    3    |  -0.229% |  -0.238% | -0.282% | +0.1%
    4    |  -0.368% |  -0.172% | -0.554% | -0.3%
    5    |  -0.274% |  -0.267% | -0.413% | -0.1%
    6    |  -0.338% |  -0.348% | -0.357% | +0.2%

STATS_CHANGED

Change-Id: I1aec9cc1a4eb06d1806d71df9dba1defad9a7e1d