Remove usage of gf_group element in LAP stage
Values of the elements of gf_group are always set
to 0 for LAP stage and hence the codeflow currently
works with that assumption. This commit puts explicit
checks at places where gf_group should not be used by
the LAP stage. This will ensure that the LAP stage
behaves correctly when the above assumption is broken.

Change-Id: I2a833497073364a5b91d7400ff82434c0919299a