Analyze per-frame noise in firstpass stats analysis.
This patch adds a new array to store the analyzed per-frame noise
variance based on the first pass statistic analysis.

Now the noise variance and the correlation coefficient of each frame
can be estimated without the knowledge of frame region types. The
logic behind this is that the noise and coeff should serve as
"second-step" firstpass statistics, and can be used in different
manners for different scenarios (e.g. for blending regions we do not
trust the accumulated correlation coefficients).

Another benefit is that now the estimated correlation coeffs are
"accumulatable" so we can better model the influence of a frame on
other frame further away.

Based on this, the patch also updated the way we decide the GOP
length. It is now not dependent on regions types (except for blendings
and scene cuts), but purely on the estimated coefficients and noise.

The overall routine is simplified. The performance remains largely
neutral. On speed 1 BD rate change (%):
            avg_psnr    ovr_psnr        ssim
lowres      -0.012      -0.112          -0.208
midres       0.014       0.065          -0.066
ugc360p     -0.158       0.064           0.049

Change-Id: If71d718c60b118f352f1fe304a07ae88c813ef72