rtc-svc: Allow for fixed SVC modes
Allow for fixed SVC modes: this is for when the
user sets the number of spatial and temporal layers but
does not use the SET_SVC_REF_FRAME_CONFIG control (flexible mode)
or the "flags" (AOM_EFLAG_) in the encode call. In such a case
the pattern is set to a fixed structure, and is only allowed
for up to 3 spatial or temporal layers. Also add ksvc_fixed_mode
flag so fixed SVC supports full SVC and KVC.

The (renamed) flag svc->set_ref_frame_config indicates the
SVC is set using the svc->(refresh/ref_idx/reference), in
either flexible or fixed mode.

For the sample enocder, svc_encoder_rtc.c, fixed mode is
tested when layering_mode = -1.

Change-Id: I84ea2c8edac393e8aa0390f1d7fdb540c183fc58