Compute gradient data once at superblock level
This CL enables the generation of gradient data at superblock level
and the consumption of the same at block level when
partition_search_type is SEARCH_PARTITION. This optimization is
enabled only for intra frames. This design avoids repetitive
calculations of gradient data at pixel level as it is calculated
based on source pixels.

For AVIF still image encode,

           Instruction Count
  speed      Reduction(%)
    0           0.863
    1           1.500
    2           1.217
    3           1.812
    4           2.844
    5           3.242
    6           0.470

BUG=aomedia:2996
BUG=aomedia:2959

Change-Id: I0cf14b50c5598654dd6419b2793908c810ec7cbe