Use a consistent tx size in first pass
The tx_size used during intra mode of first pass is inconsistent
depending on the location of the current coding block. This is a
leftover of old code from VP8, where different intra predictor is used
at different blocks during the first pass.

This commit makes the first pass to consistently use the same tx_size
TX_4X4 for the encode.

The performance is noisy, but overall there is a slight gain (~0.1%) on
hdres with small speed change (<1%).

Performance:
 SPD_SET | TESTSET | OVR_PSNR | AVG_PSNR |   SSIM  |  SPD
---------+---------+----------+----------+---------+--------
         |   HDRES |  -0.186% |  -0.112% | -0.004% | -0.0%
    1    |  MIDRES |  -0.057% |  -0.048% | +0.147% | -0.4%
         |  LOWRES |  +0.046% |  -0.007% | +0.301% | -0.6%
         |  SCREEN |  -0.018% |  -0.024% | -0.049% | +0.0%
---------+---------+----------+----------+---------+--------
         |   HDRES |  -0.057% |  -0.014% | +0.063% | -0.0%
    2    |  MIDRES |  -0.039% |  -0.027% | +0.128% | -0.5%
         |  LOWRES |  +0.003% |  -0.058% | +0.223% | -0.5%
         |  SCREEN |  -0.037% |  -0.023% | +0.019% | +0.0%
---------+---------+----------+----------+---------+--------
         |   HDRES |  -0.102% |  -0.040% | +0.015% | +0.0%
    6    |  MIDRES |  +0.057% |  +0.064% | +0.205% | -0.2%
         |  LOWRES |  -0.114% |  -0.083% | +0.058% | -0.5%
         |  SCREEN |  -0.016% |  -0.039% | -0.032% | -0.0%
---------+---------+----------+----------+---------+--------

STATS_CHANGED

Change-Id: I7592d28b389d933640524634faf623558a1724b0