Code drl context depend only on mode context
This is a step towards making the parsing dependency
in MV parsing better when NEW_INTER_MODES is on.
lowres2, midres2 (33 frames) show a loss of only 0.05
and 0.07% respectively.

Change-Id: I1544a9f5bd55a133594a31bba981c89cbd9b76b8

Remove parsing dependency on #ref_mvs

Eliminates the parsing dependency for coding drl indices
based on the number of reference MVs.

This increases the loss to about 0.2% for new_inter_modes
but improves the MV parsing dependency substantially.

Change-Id: I482fc49ffdd57864e2a1b2a2ee9616fb04ba27ab

Improve drl coding without parsing dependency

Use the refmv_count until the time mode_context is
produced to determine an estimated max #refmv_count
which is used to code the drl bits.

Reduces the loss due to no MV parsing dependency to
+0.12% from +0.25% before.

Change-Id: I2440bc00ab71e29bc654e34ce36bde7dadcada85