Removing parsing dependency for drl coding
There are several components here:
1. Make drl context depend only on mode context

This is a step towards making the parsing dependency
in MV parsing better when NEW_INTER_MODES is on.
lowres2, midres2 (33 frames) show a loss of only 0.05
and 0.07% respectively.

2. Remove parsing dependency on #ref_mvs

Eliminates the parsing dependency for coding drl indices
based on the number of reference MVs.

This increases the loss to about 0.2% for new_inter_modes
but improves the MV parsing dependency substantially.

3. Improve drl coding without parsing dependency

Use the refmv_count until the time mode_context is
produced to determine an estimated max #refmv_count
which is used to code the drl bits.

Reduces the loss due to no MV parsing dependency to
+0.12% from +0.25% before.

4. Incorporate a stronger form or parsing dep removal in NIM.

Change-Id: Id8d27957f82cd7ee95eedc08adb7fd2f10e4350c