rtc: Adjust logic in force_large_partition_blocks
For speed 9: add content metrics to logic for
increasing the partition thresholds. Avoid the increase
for superblocks that have high sad, unless the sequence
is high content one (small amount of low motion).

Small speed slowdown (~1-2%), but bdrate increase of ~4%
on some clips, with overall increase of ~0.5% on rtc set.

Change-Id: I2486e1c893fd20f95e517845ec01daefcd88e5e3