Convert input frames to monochrome before encoding
When encoding a color input to a monochrome output, convert input frames
to monochrome frames in the AOM_IMG_FMT_I420 format before passing them
to the encoder. The libaom encoder apparently has some problems with
encoding input frames in the AOM_IMG_FMT_I422 format to a monochrome
output. Although I could not get to the bottom of those problems, this
change allows us to bypass those problems.

Do not upgrade profile 0 to profile 2 when encoding to a monochrome
output, even if the input is YUV 4:2:2.

BUG=aomedia:2913

Change-Id: I86cc6a5a533092e241ba26b95d65a9cbe5fdd67f