Add a realtime mode for loop filter
Since the minimum block size and transform size of realtime mode
is 8x8, we could apply loop filtering at this granuity, instead
of 4x4.

Now it only supports single thread case.
Multithread and decoder side usage is not yet added.

With this change, the encoder speed for realtime mode is roughly
5% faster.

Change-Id: Ic1f261b75c4ee4ef234552b519fc28c155d09f07