Skip wavelet energy calculation in first pass
The stats based on wavelet energy which is calculated in
first pass is required in case of perceptual deltaq mode
and in the ML model which predicts if a flat GOP structure
is better. Thus wavelet energy calculation is enabled only
for those cases and disabled for others.

           Instruction Count
cpu-used      Reduction(%)
   2              0.138
   3              0.403
   4              0.597
   5              1.130
   6              1.459

Change-Id: I07727cbd522b0e36914c8c0f9ceaea02a2a91fc3