rtc: Some fixes to cyclic refresh
This change sets the segment on skipped blocks
such that segment_id is the same on encoder/decoder
for skip_txfm=1 blocks. Also cm->cur_frame->seg_map
needs to be updated, and over 4x4 blocks. These changes
are needed to allow for loopfilter-delta
on the segments in the cyclic refresh:
https://aomedia-review.googlesource.com/c/aom/+/135621

bdrate gain for speed 9: 0.25%, no change in speed

Change-Id: I42984f198d90f717ea5fe9a5a2ff0a8719926da5