rtc: Fixes to rate control under bandwidth change
Reset buffer level when there is significant change
in bandwidth via change_config, for both 1 layer and svc.
Also clamp the buffer_level to maximum_buffer_size in
set_rc_buffer_sizes().

This allows better response in rate control for CBR mode
when the application changes the bitrate on the fly.

Change-Id: I9ddb889c4a0ca3eb60484b127734a14b91f4c269