Add EndToEndSSIMTest to check SSIM quality based gating
The unit-tests EndToEndSSIMTest and EndToEndSSIMTestLarge are introduced
(in a new file end_to_end_ssim_test.cc) to check the adherence to a
given SSIM threshold.

The files/functions, which are needed for AOMSSIM metric calculation
have been moved outside the CONFIG_INTERNAL_STATS macro. As SSIM
related functions (which are defined in aom_dsp/ssim.c) need to
be called in end_to_end_ssim_test.cc, the file is moved under
if(NOT BUILD_SHARED_LIBS) in test.cmake.

similarity() in aom_dsp/ssim.c is also modified to silence static
analyzer warnings.

BUG=aomedia:2959

Change-Id: I6be641bfd2794511775de87a97a45a6d56101576