ERP: Fix a mismatch when ext partitions are used on the boundary
ERP had mismatch issues on the boundary of the frame. This happens when
the current block is square, so the bitstream only allows PARTITION_HORZ
or PARTITION_VERT. But the encoder might still attempt
PARTITION_HORZ/VERT_3.

This commit adds a new pruning criterion in rd_pick_partition to
determine whether the partition_3s are allowed by the syntax:

  - If the current block is 128X128, then ext partitions are not allowed.
  - Otherwise, if has_rows and has_cols are true, then ext partitions
    are allowed.
  - If has_rows and has_cols are true, but the block is square, then
    ext partitions are not allowed.

  - If ext partitions are allowed, and the current block is square, and
    chroma sizes are valid, then the partition_3s are allowed.
  - If ext partitions are allowed, and the current block is tall, then
    partition_horz_3 is allowed.
  - If ext partitions are allowed, and the current block is wide, then
    partition_vert_3 is allowed.
  - Else the partition_3's are not allowed.

Change-Id: I666896543d5e2d48f6b85a14b2e0dea24b156c42